react-renderer: Publish the reference CSS for the components
This stylesheet is put in test/ because it is not considered as part
of the codebase for now. However, it should be viewed as being highly
coupled with the components in the codebase.

The reason why the project is organized like this is that I haven't
found a good solution to write CSS in React. First, React itself
lacks of proper support for CSS selectors
(https://www.w3schools.com/cssref/css_selectors.asp), so it is
impossible to keep everything in React. Second, I don't wan't to
use CSS-in-JS tools because I want to keep the usage of this library
simple. For example, if I used styled-components, the downstream
projects such as Notablog would be forced to use it. That makes
things too complicated.

The optimal solution would be like writing CSS in Svelte, where full
CSS support is embedded in the framework, and CSS can be extracted
to a CSS string.